fix bug 507
When video is created from camera the code creates `YPMediaVideo` without asset. please check `YYPPickerVC.swift` line 78-80`. Then we YPMediaVideo after trimming and change cover in `VideoFiltersVC` line 138-139`  and it will fail if there is no asset as we are check of asset is not nil. `let asset = self?.inputVideo.asset`.