feature: Allow overriding the picker title (moving the gallery button)
Add a new configuration parameter called showsLibraryButtonInTitle,
defaulting to true (which is the original implementation). If the variable
is set to false, the Library button will be shown between the asset view
and the library view instead. In addition to the new boolean, add another
new variable (pickerTitleOverride) that will allow overriding the title
of the picker view when the showsLibraryButtonInTitle is set to false.